Sign In Functional
The sign in page is functional. It pushes the dummy home page if given valid credentials. It shows error for other scenarios (invalid creds and email not verified). This merge does not close the issue. The design is not exactly matching the requested one; but we can fix that with another merge.