API Changes & Verification Redirection
Request and response models changed to reflect the recent API changes. Email verification page now pings the server every 5 seconds to see if the user is verified, and redirects to success page automatically. The ping happens only if the app is active; it is suspended if the app is in the background or inactive (screen goes off). This will ensure the app does not keep pinging the server perpetually. Verification for regenerated OTP does not work; this is likely to be fixed when it is fixed in the backend.