Worker Initialize
When signing in as worker, if worker is not inserted, now pushes to worker details page. Theoretically, the worker can then add the details and if the insert is successful, they will be redirected to the home page. Theoretically because this has not been tested. I will test it tomorrow when I have an active owner account; so, this is not PR ready. Also, I have added a temporary logout button in the homepage to assist with testing; the logout button is functional, but it will obviously be moved and its UI changed according to the design once the homepage UI is done.